+++ addon/components/one-way-input-mask.js
@@ -1,6 +1,5 @@
-/* global Inputmask */
-
 import Component from '@ember/component';
+import Inputmask from 'inputmask';
 import { computed, get, set } from '@ember/object';
 import { schedule } from '@ember/runloop';
 import { areDifferent } from 'ember-inputmask/utils/compare-objects';
@@ -71,7 +70,7 @@ const OneWayInputMask = Component.extend({
    * @argument value
    * @type String
    */
-  value: null,
+  value: '',
 
   /**
    * Setup _value to be a positional param or the passed param if that is not defined
@@ -157,7 +156,7 @@ const OneWayInputMask = Component.extend({
    * @argument update
    * @type function
    */
-  update() {},
+  update: null,
 
   /**
    * _changeEventListener - A place to store the event listener we setup to listen to the 'input'