+++ tests/dummy/app/components/one-way-currency-mask/demo-3/component.js
@@ -1,5 +1,6 @@
 import Component from '@glimmer/component';
+import { tracked } from '@glimmer/tracking';
 
 export default class Demo3Component extends Component {
-
+  @tracked foo;
 }