+++ tests/dummy/app/components/one-way-email-mask/demo-1/component.js
@@ -1,5 +1,6 @@
 import Component from '@glimmer/component';
+import { tracked } from '@glimmer/tracking';
 
 export default class EmailMaskComponent extends Component {
-
+  @tracked foo;
 }