+++ tests/dummy/app/components/one-way-input-mask/demo-1/component.js
@@ -1,0 +1,8 @@
+import Component from '@glimmer/component';
+import { tracked } from '@glimmer/tracking';
+
+// BEGIN-SNIPPET one-way-input-mask-demo-1.js
+export default class OneWayInputMaskDemo1Component extends Component {
+  @tracked foo;
+}
+// END-SNIPPET