+++ tests/dummy/app/components/one-way-zip-code-mask/demo-2/component.js
@@ -1,4 +1,6 @@
 import Component from '@glimmer/component';
+import { tracked } from '@glimmer/tracking';
 
 export default class Demo2Component extends Component {
+  @tracked foo;
 }